src/main/python/weather_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-service"
version = "1.0.0"
description = "Weather forecast service - Meteosource API to InfluxDB/MQTT"
license = {text = "MIT"}
authors = [
    {name = "Robert Fiser"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Home Automation",
]
dependencies = [
    "paho-mqtt>=1.6.0,<3.0.0",
    "requests>=2.28.0",
    "influxdb-client>=1.36.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
weather-service = "weather_service:main"

[tool.setuptools]
packages = ["weather_service"]
